operator: move code to convert to legacy config to a new package
Let's add a new package to deal with conversions from legacy
configuration format to the new one: the 'converter' pkg.
Let's move out the code from the server package.
We will soon need to do the same for the register client side, sharing
the LegacyConfig struct and isolating the code in a separate package
would make it more manageable.
